Split up the Timetable page into two "stages" as before
In the original VAtriume app, the Timetable service was split into two
stages: course selection and section selection with drag-n-drop. In this
version, I wanted to come up with something different, but select->arrange
seems to be the best option.

As for why, the course selection process might be later enhanced by suggestions
based on the student's current year of study, remaining credits, and other
helpful metrics. For instance, CS students usually take a DBMS class in the
3rd year of their studies, hence this course should be at the top of search
results.

This also leaves more screen real estate to the timetable itself, which is
vital.